Fix project tests

+++ test/JSONEndPoints/ToDo/UndocumentedToDoEndPoints/ToDoProjectRelationshipEndPoint_test.py
@@ -7,7 +7,7 @@ import psutil
 BASE_URL = "http://localhost:4567"
 TODOS_ENDPOINT = "/todos"
 TODO_PROJ_RELATIONSHIP = "tasksof"
-JAR_PATH = "../../../../runTodoManagerRestAPI-1.5.5.jar"
+JAR_PATH = "../../../runTodoManagerRestAPI-1.5.5.jar"
 
 @pytest.fixture(scope="function", autouse=True)
 def setup_and_teardown():
@@ -42,11 +42,16 @@ def setup_and_teardown():
         print("Server did not respond to shutdown request.")
 
     # Ensure the Java process is killed
-    parent = psutil.Process(process.pid)
-    for child in parent.children(recursive=True):  # Kill child processes
-        child.terminate()
-    process.terminate()
-    process.wait()
+    try:
+        parent = psutil.Process(process.pid)
+        for child in parent.children(recursive=True):  # Kill child processes
+            if child.is_running():
+                child.terminate()
+        if parent.is_running():
+            parent.terminate()
+        parent.wait()
+    except psutil.NoSuchProcess:
+        pass
 
 def test_get_projects_for_todos():
     expected = {
